pyproject.toml
[tool.poetry]
name = "qr_code_generator"
version = "0.1.0"
description = ""
authors = ["Gobot1234 <gobot1234yt@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26.2"
Pillow = "^10.1.0"
qrcode = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
